`callRenderer` to `respondToRenderer`
